Task #4704: Merged latest trunk into the branch and fixed compilation errors. Four tests (still) fail.

+++ branches/Cobalt-Task4704/MAC/MACIO/include/MACIO/StringVector.h
@@ -49,9 +49,9 @@ public:
 	StringVector() {};
 	~StringVector() {};
 
-	size_t getSize() const;
-	size_t pack  (char	*buffer) const;
-	size_t unpack(char	*buffer);
+	uint32 getSize() const;
+	uint32 pack  (char	*buffer) const;
+	uint32 unpack(char	*buffer);
 	ostream& print (ostream& os) const;
 
 	vector<string>&	operator()()
@@ -66,23 +66,23 @@ private:
 // @}
 
 // getSize()
-inline size_t StringVector::getSize() const
+inline uint32 StringVector::getSize() const
 {
 	return (MSH_size(theVector));
 }
 
 // pack()
-inline size_t StringVector::pack(char	*buffer) const
+inline uint32 StringVector::pack(char	*buffer) const
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 	MSH_pack(buffer, offset, theVector);
 	return (offset);
 }
 
 // unpack()
-inline size_t StringVector::unpack(char	*buffer)
+inline uint32 StringVector::unpack(char	*buffer)
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 	MSH_unpack(buffer, offset, theVector);
 	return (offset);
 }
